revert deprecation of device flags
