+++ pytorch_lightning/trainer/trainer.py
@@ -2084,10 +2084,6 @@ class Trainer(
 
     @property
     def num_processes(self) -> int:
-        rank_zero_deprecation(
-            "`Trainer.num_processes` is deprecated in v1.6 and will be removed in v1.8. "
-            "Please use `Trainer.num_devices` instead."
-        )
         return self.num_devices
 
     @property
@@ -2104,10 +2100,6 @@ class Trainer(
 
     @property
     def ipus(self) -> int:
-        rank_zero_deprecation(
-            "`Trainer.ipus` was deprecated in v1.6 and will be removed in v1.8."
-            " Please use `Trainer.num_devices` instead."
-        )
         return self.num_devices if isinstance(self.accelerator, IPUAccelerator) else 0
 
     @property
@@ -2195,10 +2187,6 @@ class Trainer(
 
     @property
     def gpus(self) -> Optional[Union[List[int], str, int]]:
-        rank_zero_deprecation(
-            "`Trainer.gpus` was deprecated in v1.6 and will be removed in v1.8."
-            " Please use `Trainer.num_devices` or `Trainer.device_ids` to get device information instead."
-        )
         return self._accelerator_connector.gpus
 
     @property
